md_Destructive_LAI/requirements.txt
setuptools==44.0.0
wheel==0.45.1
click==8.1.7
numpy==1.24.4
matplotlib==3.7.5
pandas==2.0.3
scikit-learn==1.3.2
geopandas==0.13.2
rasterio==1.3.11
pillow==10.4.0
laspy==2.5.4
shapely==2.0.7
fiona==1.9.0
qrcode==7.4.2
pyyaml==6.0.2
tqdm==4.67.1
pyzbar==0.1.9
PyQt5==5.15.11
opencv-python-headless==4.11.0.86
blend_modes==2.2.0
open3d==0.19.0
torch
